requirements.txt
alabaster==0.7.10
appdirs==1.4.0
astroid==1.4.9
Babel==2.4.0
bleach==1.5.0
click==6.7
ConfigArgParse==0.10.0
coverage==4.4.1
cycler==0.10.0
Cython==0.23.1
decorator==4.0.11
docutils==0.13.1
entrypoints==0.2.2
flake8==3.0.4
Flask==0.12.2
funcsigs==1.0.2
future==0.15.2
h5py==2.6.0
hmmlearn==0.2.1
html5lib==0.999
imagesize==0.7.1
ipykernel==4.5.2
ipython==5.3.0
ipython-genutils==0.1.0
ipywidgets==6.0.0
isort==4.2.5
itsdangerous==0.24
Jinja2==2.9.5
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==5.0.0
jupyter-console==5.1.0
jupyter-core==4.3.0
Keras==2.0.2
lazy-object-proxy==1.3.1
MarkupSafe==0.23
matplotlib==2.0.0
mccabe==0.5.3
mistune==0.7.4
nbconvert==5.1.1
nbformat==4.3.0
neon==1.9.0
notebook==4.4.1
numpy==1.12.0
olefile==0.44
opencv-python==3.2.0.6
pandas==0.19.2
pandocfilters==1.4.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.3.1
posix-ipc==1.0.0
prompt-toolkit==1.0.13
protobuf==3.3.0
ptyprocess==0.5.1
py==1.4.32
pycodestyle==2.0.0
pyflakes==1.2.3
Pygments==2.2.0
pylint==1.6.4
pyparsing==2.1.4
pytest==3.0.1
pytest-cov==2.3.1
pytest-mock==1.6.0
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.2.1
requests==2.14.2
requests-toolbelt==0.7.1
scikit-learn==0.18.1
scipy==0.18.1
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.4.6
tensorflow==1.0.0
terminado==0.6
testpath==0.3
Theano==0.9.0
tornado==4.4.2
tqdm==4.8.4
traitlets==4.3.2
udacity-pa==0.2.4
wcwidth==0.1.7
Werkzeug==0.12.2
widgetsnbextension==2.0.0
wrapt==1.10.10
